# contracts/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API behavior, compatibility rules, and endpoint contract requirements.",
    "version": "1.25.1"
  },
  "paths": {
...
    "/api/spool/{namespace}/{query_id}.parquet": {
      "get": {
        "summary": "GET /api/spool/{namespace}/{query_id}.parquet",
        "responses": {
          "200": {
            "description": "Contract response: application/octet-stream (parquet)"
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "410": {
            "description": "Error response (see contract error format)"
          }
        },
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-response-contract": "application/octet-stream (parquet)",
        "x-cdd-errors": "400/410"
      }
    },
...
    "/api/resource/status": {
      "get": {
        "summary": "GET /api/resource/status",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "500": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "400/500"
      }
    },
...
    "/api/resource/history/options": {
      "get": {
        "summary": "GET /api/resource/history/options",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "500"
      }
    },
    "/api/resource/history/query": {
      "post": {
        "summary": "POST /api/resource/history/query",
        "responses": {
          "201": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "202": {
            "description": "Error response (see contract error format)"
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "410": {
            "description": "Error response (see contract error format)"
          },
          "500": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Contract request: JSON body (schema not machine-resolved; see contract)",
          "content": {},
          "x-cdd-unresolved": true
        },
        "x-cdd-errors": "202/400/410/500"
      }
    },
    "/api/resource/history/view": {
      "get": {
        "summary": "GET /api/resource/history/view",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "410": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "400/410"
      }
    },
    "/api/resource/history/page": {
      "get": {
        "summary": "GET /api/resource/history/page",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "—"
      }
    },
    "/api/resource/history/export": {
      "get": {
        "summary": "GET /api/resource/history/export",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "410": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "400/410"
      },
      "post": {
        "summary": "POST /api/resource/history/export",
        "responses": {
          "201": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "410": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Contract request: JSON body (schema not machine-resolved; see contract)",
          "content": {},
          "x-cdd-unresolved": true
        },
        "x-cdd-errors": "400/410"
      }
    },
    "/api/resource/history/query/progress": {
      "get": {
        "summary": "GET /api/resource/history/query/progress",
        "responses": {
          "200": {
            "description": "Contract response: ProgressResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "404": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "400/404"
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "GenericSuccessResponse": {
        "type": [
          "object",
          "null"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {},
          "error": {},
          "meta": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string"
              },
              "app_version": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ProgressResponse": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "query_id": {
                "type": "string"
              },
              "total_chunks": {
                "type": "integer"
              },
              "completed_chunks": {
                "type": "integer"
              },
              "percent": {
                "type": "number"
              },
              "status": {
                "type": "string"
              }
            }
          },
          "meta": {
            "type": "object",
            "required": [
              "timestamp"
            ],
            "properties": {
              "timestamp": {
                "type": "string"
              },
              "app_version": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  }
}